Save my workup of alternate footprints for kicad.
We're probably going to move forward on a bunch of this, keeping our
own.  We're keeping lots of inspiration from KLC, but not being so rigid
with it.

First off, a bunch of dual LEDs I needed to compare for subs.  To make
it easier to compare and find, they have a different naming structure.

T for top, B for bottom
xxyy metric dimensions
_VENDOR_
_MODEL_PREFIX_

LED_Dual is for 4 terminal _dual_ leds, all with pins renumbered to be
the same style, _regardless_ of what the vendor datasheet says!  This is
the biggest departure from the KLC, but allows _vastly_ easier
comparison when looking for alternate parts.

3d model here is from https://grabcad.com/library/led-s-smd-single-bi-and-tri-color-1